moved host zone query test
